Rename create_CC_mesh_from_contours to CCMesh.from_contours and make it a class method Remove vox2ras_tkr from fsaverage information Fix surface saving Fix --slice_selection type function Fix mapping from upright to orig space Change types of ndarrays from python to numpy Scalar types

+++ CorpusCallosum/shape/contour.py
@@ -73,8 +73,8 @@ class CCContour:
     
     def __init__(
         self,
-        contour: np.ndarray[tuple[Literal["N", 2]], np.dtype[float]],
-        thickness_values: np.ndarray[tuple[Literal["N"]], np.dtype[float]],
+        contour: np.ndarray[tuple[Literal["N", 2]], np.dtype[np.float_]],
+        thickness_values: np.ndarray[tuple[Literal["N"]], np.dtype[np.float_]],
         endpoint_idxs: tuple[int, int] | None = None,
         resolution: float = 1.0
     ):
